inference/requirements.txt
torch>=2.0.0
transformers>=4.30.0
peft>=0.4.0
cadquery>=2.4.0
numpy-stl>=2.16.0
openai>=1.0.0
Pillow>=9.5.0
tqdm>=4.65.0
numpy>=1.24.0
matplotlib>=3.7.0